database script updated

+++ playlist_generator/src/main/java/track_ninja/playlist_generator/repositories/AuthorityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import track_ninja.playlist_generator.models.Authority;
 
 @Repository
-public interface AuthorityRepository extends CrudRepository<Authority, Long> {
+public interface AuthorityRepository extends CrudRepository<Authority, Integer> {
 }
