+++ playlist_generator/src/main/java/track_ninja/playlist_generator/repositories/PlaylistRepository.java
@@ -7,7 +7,7 @@ import track_ninja.playlist_generator.models.Playlist;
 import java.util.List;
 
 @Repository
-public interface PlaylistRepository extends CrudRepository<Playlist, Long> {
+public interface PlaylistRepository extends CrudRepository<Playlist, Integer> {
     List<Playlist> findPlaylistsByGenres_name(String genre);
 
     List<Playlist> findAllByUser_User_Username(String username);
